# backend/src/main/java/com/trodriguez/users/users/dto/UserDTO.java
package com.trodriguez.users.users.dto;

import java.util.Date;
import java.util.UUID;

public record UserDTO(
    UUID id,
    String firstNames,
    String lastNames,
    Long rutNumber,
    String verificationDigit,
    Date birthDate,
    String email
) {}